Group Commit. Login blueprint changed to Auth. Moved the actions worked on in the genre model to CrudMixin to enable use in all models. Added session setting and control. Added passport to user model, this will be used to validate the user further (prevent session user_id guessing)

+++ app/__init__.py
@@ -1,9 +1,6 @@
-from flask import Flask
-from flask_bigapp import BigApp
-from flask_sqlalchemy import SQLAlchemy
+from pathlib import Path
 
-bigapp = BigApp()
-db = SQLAlchemy()
+from app.extensions import *
 
 
 def create_app():
@@ -13,6 +10,10 @@ def create_app():
         template_folder="templates",
         static_url_path="/static"
     )
+
+    sysconf_location = Path(Path(app.root_path) / "sysconf.ini")
+    sysconf.read(Path(sysconf_location))
+
     bigapp.init_app(app)
     db.init_app(app)
 
@@ -21,38 +22,42 @@ def create_app():
     bigapp.import_blueprints("blueprints")
 
     with app.app_context():
-        db.drop_all()
-        db.create_all()
-        add_result = bigapp.model("Genre").create(
-            batch=[
-                {"genre": "Action"},
-                {"genre": "Horror"},
-                {"genre": "Adventure"},
-                {"genre": "Comedy"},
-                {"genre": "Crime"},
-                {"genre": "Drama"},
-                {"genre": "Fantasy"},
-                {"genre": "Historical"},
-            ]
-        )
-        # result = bigapp.model("Genre").create(
-        #     genre="Horror",
-        # )
-        print("add_result")
-        print(add_result)
+        if sysconf.getboolean("SYSTEM", "FIRST_RUN"):
+            print("First run detected, creating database...")
+            sysconf.set("SYSTEM", "FIRST_RUN", "false")
+            with open(sysconf_location, "w") as f:
+                sysconf.write(f)
 
-        view_result = bigapp.model("Genre").read(id_=1)
-        print("view_result")
-        print(view_result)
+            db.drop_all()
+            db.create_all()
+            bigapp.model("Genre").create(
+                batch=[
+                    {"genre": "Mythical Historical Fantasy", "description": "Swords, shields, bows and magic"},
+                    {"genre": "Cult Mystery", "description": "Cults, mysteries and conspiracies"},
+                    {"genre": "Zombie Apocalypse", "description": "Zombies, bandits and looting"},
+                    {"genre": "Apocalypse", "description": "Bandits and looting"},
+                    {"genre": "Crime Film Noir", "description": "Smoke and back alleys, see!"},
+                    {"genre": "Existential Horror", "description": "Your existence means nothing."},
+                    {"genre": "Dystopian Future", "description": "The future is bleak"},
+                    {"genre": "Mythic Folklore", "description": "Werewolves, vampires, ghosts and other monsters"},
+                    {"genre": "Science Fiction", "description": "Aliens, robots, space travel and other cool stuff"},
+                ]
+            )
 
-        update_result = bigapp.model("Genre").update(
-            field=("genre", "Horror"), values={"description": "Horror"}
-        )
-        print("update_result")
-        print(update_result)
+            default_admin_email_address = "admin@users.system"
+            default_admin_password = "adminpassword"
+            passport = auth.generate_numeric_validator(6)
+            salt = auth.generate_salt()
+            password = auth.sha_password(default_admin_password, salt)
 
-        delete_result = bigapp.model("Genre").delete(id_=1)
-        print("delete_result")
-        print(delete_result)
+            bigapp.model("User").create(
+                fields={
+                    "email_address": default_admin_email_address,
+                    "password": password,
+                    "salt": salt,
+                    "passport": passport,
+                    "user_type": 10,
+                }
+            )
 
     return app
